Fix spelling of "Infrastructure" across the project
Updated project structure to correct "Infrastucture" to "Infrastructure" in all relevant files, namespaces, and references. Adjusted project references in `.csproj` files and modified the Dockerfile to ensure correct paths. Updated `packages.lock.json` to reflect changes in project structure and dependencies. Ensured consistency in tests and related files with the corrected namespace.
