+++ src/Infrastructure.SqlServer/Data/ApplicationDbContextDesignTimeFactory.cs
@@ -4,9 +4,9 @@ using System.Reflection;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
 
-using MoneyGroup.Infrastucture.Data;
+using MoneyGroup.Infrastructure.Data;
 
-namespace MoneyGroup.Infrastucture.PostgreSql.Data;
+namespace MoneyGroup.Infrastructure.SqlServer.Data;
 
 [ExcludeFromCodeCoverage]
 internal class ApplicationDbContextDesignTimeFactory : IDesignTimeDbContextFactory<ApplicationDbContext>
@@ -14,7 +14,7 @@ internal class ApplicationDbContextDesignTimeFactory : IDesignTimeDbContextFacto
     public ApplicationDbContext CreateDbContext(string[] args)
     {
         var options = new DbContextOptionsBuilder<ApplicationDbContext>()
-            .UseNpgsql(opt => opt.MigrationsAssembly(Assembly.GetExecutingAssembly().FullName))
+            .UseSqlServer(opt => opt.MigrationsAssembly(Assembly.GetExecutingAssembly().FullName))
             .Options;
 
         return new ApplicationDbContext(options);
