+++ src/Infrastructure/MoneyGroup.Infrastructure.csproj
