+++ test/IntegrationTests/Data/OrderRepositoryTest.cs
@@ -2,7 +2,7 @@
 
 using MoneyGroup.Core.Entities;
 using MoneyGroup.Core.Models.Orders;
-using MoneyGroup.Infrastucture.Data;
+using MoneyGroup.Infrastructure.Data;
 using MoneyGroup.IntegrationTests.Fixtures;
 
 namespace MoneyGroup.IntegrationTests.Data;
